Update HAJ structure
