+++ src/app/pages/History/index.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import MainLayout from "app/layouts/MainLayout";
+
+const History: React.FC<Props> = () => {
+  return (
+    <MainLayout>
+      <p>History</p>
+    </MainLayout>
+  );
+};
+
+export default History;
+
+interface Props {}
