+++ src/app/pages/Homepage/components/SupportCenter.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Theme } from "@mui/material";
+import { makeStyles } from "@mui/styles";
+
+const SupportCenter: React.FC<Props> = () => {
+  const classes = useStyles();
+
+  return <p>Fixed link area</p>;
+};
+
+export default SupportCenter;
+
+const useStyles = makeStyles((theme: Theme) => ({}));
+
+interface Props {}
