+++ src/app/pages/Jobs/components/ToggleUISwitcher.tsx
