+++ src/app/pages/Setting/index.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import MainLayout from "app/layouts/MainLayout";
+
+const Setting: React.FC<Props> = () => {
+  return (
+    <MainLayout>
+      <p>Setting</p>
+    </MainLayout>
+  );
+};
+
+export default Setting;
+
+interface Props {}
